requirements.txt
async_timeout==3.0.1
discord.py==1.2.5
youtube_dl==2020.1.15
discord==1.0.1
python-dotenv==0.10.5
